+++ src/Logary/Internals/InternalLogger.fs
@@ -1,6 +1,5 @@
 namespace Logary.Internals
 
-open System
 open Hopac
 open Hopac.Infixes
 open Logary
@@ -8,7 +7,7 @@ open Logary.Internals
 open Hopac.Extensions
 
 [<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
-module InternalLogger =
+module internal InternalLogger =
   /// This logger is special: in the above case the Registry takes the responsibility
   /// of shutting down all targets, but this is a stand-alone logger that is used
   /// to log everything in Logary with, so it needs to capable of handling its