+++ src/tests/Logary.Targets.InfluxDb.Tests/Target.fs
@@ -20,8 +20,8 @@ open Hopac.Infixes
 let start (port: int) =
   let uri = Uri (sprintf "http://127.0.0.1:%i/write" port)
   let targConf = InfluxDbConf.create(uri, "tests")
-  emptyRuntime >>= (fun (ri, ilogger) ->
-  Target.create ri (create targConf "influxdb"))
+  emptyRuntime >>= fun ri ->
+  Target.create ri (create targConf "influxdb")
 
 let shutdown t =
   job {